feat(frontend): update entrypoint form
This commit makes the following improvements to the entrypoint form:
- replace the static param form elements with dialog box forms
- make adding entrypoint and artifact params work as adding a new row to the tables.
- move entrypoint and artifact params tables to be side-by-side
- remove paging from entrypoint param and artifact param tables
- change instances of the text "params" to "parameters"

feat: turn artifact param form into dialog
