+++ src/frontend/src/components/TableComponent.vue
@@ -17,6 +17,7 @@
     :tabindex="props.disableSelect ? '' : '0'"
     @keydown="keydown"
     :rows-per-page-options="props.showAll ? [0] : [5,10,15,20,25,50,0]"
+    :hideBottom="props.hideBottom && rows.length > 0"
   >
     <template v-slot:header="props">
       <q-tr :props="props">
@@ -172,6 +173,7 @@
   disableSelect: Boolean,
   hideOpenBtn: Boolean,
   hideDeleteBtn: Boolean,
+  hideBottom: Boolean,
   rightCaption: String,
   showAll: Boolean,
   highlightRow: Boolean,
